Conflicts:
	classifier/Classifier.cc
	classifier/Main.cc
	coprocessor/CMakeLists.txt
	coprocessor/Coprocessor.cc
	coprocessor/CoprocessorTypes.h
	coprocessor/Main.cc
	coprocessor/OutputFormula.cc
	coprocessor/techniques/BVA.cc
	coprocessor/techniques/BVE.cc
	coprocessor/techniques/BVEParallel.cc
	coprocessor/techniques/EquivalenceElimination.cc
	coprocessor/techniques/FourierMotzkin.cc
	coprocessor/techniques/Probing.cc
	coprocessor/techniques/Propagation.h
	coprocessor/techniques/Rewriter.cc
	coprocessor/techniques/Subsumption.cc
	coprocessor/techniques/Unhiding.cc
	mprocessor/Main.cc
	pcasso/Main.cc
	pcasso/Master.cc
	pcasso/SolverPT.cc
	pfolio/Main.cc
	qprocessor/Main.cc
	riss/core/Main.cc
	riss/core/Solver.cc
	riss/mtl/Heap.h
	riss/mtl/Vec.h
	riss/simp/Main.cc

+++ pcasso/LookaheadSplitting.h
@@ -27,7 +27,7 @@ class LookaheadSplitting : public SplitterSolver
     // more global data structures
     Riss::vec<Riss::Lit> learnt_clause;
     Riss::vec<Riss::CRef> otfssClauses;
-    uint64_t extraInfo;
+    unsigned extraInfo;
 
   public:
     LookaheadSplitting(Riss::CoreConfig* config);
